Purged all support for runtime resolved addresses in the instruction stream. Instead we only allow runtime resolved symbol ID values. There are three classes of symbol ID:  * Data Symbol  * Code Symbol  * Native Code Symbol
Each ID is an unsigned 16-bit integer in the instruction stream after the call/load/store operation that references it. An array of
Symbol ID to host native pointer is setup within the VM.

An error is raised if any symbol ID exceeds the known range of symbol ID for the associated type.